requirements.txt
amqp==2.6.1
celery==4.4.7
coreapi==2.3.3
dicttoxml==1.7.4
Django==3.1.5
django-admin-rangefilter==0.7.0
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.1
django-filter==2.4.0
django-model-utils==4.1.1
django-redis==4.12.1
django-rest-swagger==2.2.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
flower==0.9.7
gunicorn==20.0.4
Pillow==8.1.2
psycopg2==2.8.6
PyJWT==2.0.1
pyotp==2.6.0
python-crontab==2.5.1
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
sentry-sdk==1.0.0
simplejson==3.16.0
sqlparse==0.4.1
urllib3==1.26.3
xmltodict==0.12.0
drf-spectacular==0.16.0
whitenoise==5.2.0